+++ packages/hono-backend/src/common/errors.ts
@@ -20,6 +20,14 @@ export interface AppErrorOptions {
     description?: string
 }
 
+/**
+ * Use AppError at service or route boundaries when an error should become a
+ * controlled API response with a stable HTTP status and internal code.
+ *
+ * Lower-level code should prefer regular typed domain errors, such as
+ * NoPathFoundError or StationNotFoundError, and let the service layer translate
+ * them into AppError when crossing into API-facing behavior.
+ */
 export class AppError extends Error {
     public readonly statusCode: ContentfulStatusCode
     public readonly internalCode: InternalCode
@@ -31,5 +39,22 @@ export class AppError extends Error {
         this.statusCode = statusCode
         this.internalCode = internalCode
         this.description = description
+    }
+}
+
+export class NoPathFoundError extends Error {
+    constructor(
+        public readonly from: string,
+        public readonly to: string
+    ) {
+        super('No path found between stations')
+        this.name = 'NoPathFoundError'
+    }
+}
+
+export class StationNotFoundError extends Error {
+    constructor(public readonly stationId: string) {
+        super('Station not found')
+        this.name = 'StationNotFoundError'
     }
 }
